pyproject.toml
[project]
name = "pydatagraphs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
addopts = "-ra -q"
testpaths = [
    "tests",
    "tests_integration"
]
